modelo/ml-recomendator/requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
pyyaml
joblib
httpx
python-dotenv
fastapi
uvicorn
